swap custom SeriesList for InfluxDB models.row
